renamed signup form to follow convention
