+++ src/pages/signup/__tests__/signup.js
@@ -1,9 +1,9 @@
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import React from 'react';
-import { SignUp } from '../singup';
+import { SignUp } from '../SignUp';
 
-describe('SignInForm test', () => {
+describe('SignUpForm test', () => {
   it('renders without crashing', () => {
     const { getByText } = render(<SignUp />);
     expect(getByText('Email')).toBeInTheDocument();
